git-svn-id: https://svn.apache.org/repos/asf/sis/branches/JDK6@1551742 13f79535-47bb-0310-9956-ffa450edef68

+++ core/sis-utility/src/main/java/org/apache/sis/math/StatisticsFormat.java
@@ -122,6 +122,8 @@ public class StatisticsFormat extends TabularFormat<Statistics> {
 
     /**
      * Returns the kind of objects formatted by this class.
+     *
+     * @return {@code Statistics.class}
      */
     @Override
     public Class<Statistics> getValueType() {
@@ -184,6 +186,9 @@ public class StatisticsFormat extends TabularFormat<Statistics> {
 
     /**
      * Not yet implemented.
+     *
+     * @return Currently never return.
+     * @throws ParseException Currently never thrown.
      */
     @Override
     public Statistics parse(CharSequence text, ParsePosition pos) throws ParseException {
@@ -355,7 +360,7 @@ public class StatisticsFormat extends TabularFormat<Statistics> {
      * switch statements inside the {@link #format(Statistics[], Appendable)} method
      * (we define this static field close to the format methods for this purpose).
      */
-    private static final int[] KEYS = {
+    private static final short[] KEYS = {
         Vocabulary.Keys.NumberOfValues,
         Vocabulary.Keys.NumberOfNaN,
         Vocabulary.Keys.MinimumValue,
